% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.fn.R
\name{cjs.fn}
\alias{cjs.fn}
\title{Summarize detection histories provide CJS estimates}
\usage{
cjs.fn(detect.in, L.in = NULL, seeds.in = NULL, se.out = F, d = NULL)
}
\arguments{
\item{detect.in}{Detection history in wide ATlAS format, w/o rel.group or bin columns}

\item{L.in}{L.in: vector of mean probabilities tag is working, given detection at each site. If NULL function will estimate the CJS parameters w/o tag failure.}

\item{seeds.in}{optional starting estimates for CJS model parameters Si, pi, li where i is 1:(total number of sites-1)}

\item{se.out}{calculate se(cjs) from variance matrix of parameters using Hessian? Logical.}

\item{d}{proportion of detected tags censored at a location}
}
\value{
2-column matrix with CJS estimates and standard errors (if estimated).
}
\description{
CJS.fn summarizes the individual detection histories, estimates seeds for Cormack-Jolly-Seber likelhood fn
}
