% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglife.fn.R
\name{fail.fn}
\alias{fail.fn}
\title{Predicted survival fraction for 3-parameter weibull distribution}
\usage{
fail.fn(x, b.in, g.in, n.in)
}
\arguments{
\item{x}{time}

\item{b.in}{beta}

\item{g.in}{gamma}

\item{n.in}{eta}
}
\value{
vector of proportion active at time x
}
\description{
Predicted survival fraction for 3-parameter weibull distribution
}
