% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic}
\alias{harmonic}
\title{Estimate harmonic mean}
\usage{
harmonic(t.in)
}
\arguments{
\item{t.in}{numeric vector (no NA's)}
}
\value{
character vector of harmonic mean and se
}
\description{
Estimate harmonic mean
}
