% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglife.fn.R
\name{logweib.lik}
\alias{logweib.lik}
\title{Title}
\usage{
logweib.lik(x, tags.in)
}
\arguments{
\item{x}{Log-likelihood of 3-parameter weibull distribution (beta, gamma, eta)}

\item{tags.in}{observed time to failure}
}
\value{
negative log-likelihood
}
\description{
Title
}
