% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.tt2site.fn.R
\name{mean.tt2site.fn}
\alias{mean.tt2site.fn}
\title{Estimate the travel time harmonic mean, release to each detection site}
\usage{
\method{mean}{tt2site.fn}(data.in, num.period, site.names)
}
\arguments{
\item{data.in}{detection history and detection times (flat format)}

\item{num.period}{number of detection sites}

\item{site.names}{vector of site names}
}
\value{
harmonic.tt: harmonic mean of travel time (days); tt.matrix: time of release to 1st detection time for each site (days); activetime.matrix: time of tag activation to 1st detection for each site (days)
}
\description{
Estimate the travel time harmonic mean, release to each detection site
}
