% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mttf.R
\name{mttf}
\alias{mttf}
\title{Estimate mean time to failure based on 3-parameter Weibull}
\usage{
mttf(x)
}
\arguments{
\item{x}{vector of beta, gamma, eta parameters of Weibull distribution}
}
\value{
mean and sd of time to failure from Weibull curve
}
\description{
Estimate mean time to failure based on 3-parameter Weibull
}
