% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single.rel}
\alias{single.rel}
\title{Single Release ATLAS Practice File.}
\format{
A data frame with 8 columns and 9996 rows in the ATLAS file format:
\describe{
\item{V1}{Release group (ie. R1)}
\item{V2}{tag lot number (ie. 1,2,...)}
\item{V3}{tag id code}
\item{V4}{tag activation date/time  (yyyy-mm-dd hh:mm:ss)}
\item{V5}{release date/time  (yyyy-mm-dd hh:mm:ss)}
\item{V6}{detecton site designation}
\item{V7}{detection indicator (0: not detected; 1: detected; 2: detected & censored)}
\item{V8}{tag detection date/time  (yyyy-mm-dd hh:mm:ss) at that site}
...
}
}
\usage{
single.rel
}
\description{
Single Release ATLAS Practice File.
}
\keyword{datasets}
