% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taglife.data}
\alias{taglife.data}
\title{Taglife ATLAS Practice File.}
\format{
Taglife study results with length of time to failure for 100 tags.  Tags were
observed continually until failure.  A data frame with 1 columns and 100 rows:
\describe{
\item{tag_life_days}{length of time from activation to tag failure in days.}
}
}
\usage{
taglife.data
}
\description{
Taglife ATLAS Practice File.
}
\references{
\href{https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=&cad=rja&uact=8&ved=2ahUKEwjv262c3Ln2AhW2JzQIHde_B40QFnoECAgQAQ&url=http\%3A\%2F\%2Fpweb.crohms.org\%2Ftmt\%2Fdocuments\%2FFPOM\%2F2010\%2FNWW\%2520Research\%2FCompliance\%2520Monitoring\%2520at\%2520McNary\%2520Dam\%2520in\%25202014_Final\%2520Report.pdf&usg=AOvVaw1LD6qSRn8oChimStypCy57}{McNary Dam survival study}
}
\keyword{datasets}
