% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglife.fn.R
\name{taglife.fn}
\alias{taglife.fn}
\title{Fitting tag-life function}
\usage{
taglife.fn(tags.in, model.in = "weibull", tag.se = T, plot.taglife)
}
\arguments{
\item{tags.in}{vector of observed time to failure (days)}

\item{model.in}{name of model to use.  Current option is "weibull"}

\item{tag.se}{compute SEs of estimates}

\item{plot.taglife}{create figure to see how model fits}
}
\value{
List with MLE estimates, mean time to failure based on selected model, andvector of tags used in the estimation of the tag-life curve.
}
\description{
Fitting tag-life function
}
