% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.fn.R
\name{thist0}
\alias{thist0}
\title{Title}
\usage{
thist0(x)
}
\arguments{
\item{x}{detection history in wide ATlAS format, w/o rel.group or bin columns}
}
\value{
summary of counts.in/history and unique detection matrix in same order
}
\description{
Title
}
