% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas2flat.fn.R
\name{atlas2flat.fn}
\alias{atlas2flat.fn}
\title{Title}
\usage{
atlas2flat.fn(data.in)
}
\arguments{
\item{data.in}{table: columns = release group name, bin number, tag id, tag activation date/time, tag release date/time, site name, detection (1=yes, 0 = no), detection date/time}
}
\value{
table: columns = release group name, bin number, tag id, activation date/time, tag release date/time, 1 column per site name with detection time (blank if no detection)
}
\description{
Title
}
