% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.fn.R
\name{cjs.fn}
\alias{cjs.fn}
\title{Title}
\usage{
cjs.fn(
  detect.in,
  L.in = NULL,
  seeds.in = NULL,
  se.out = F,
  d = NULL,
  bootstrap = F
)
}
\arguments{
\item{detect.in}{Detection history in wide ATlAS format, w/o rel.group or bin columns}

\item{L.in}{L.in: vector of mean probabilities tag is working, given detection at each site. If NULL function will estimate the CJS parameters w/o tag failure.}

\item{seeds.in}{seeds.in: optional starting estimates for CJS model, includes additional p(censoring) per site at end parameters are Si, pi, li where i is 1:(total number of sites-1) di (p(censoring)), is not estimated in this version}

\item{se.out}{calculate se(cjs var) from variance matrix of parameters using Hessian? Logical.}

\item{d}{???}

\item{bootstrap}{Bootstrap or not. Logical.}
}
\value{

}
\description{
Title
}
