% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.fn.R
\name{cjs.fn}
\alias{cjs.fn}
\title{Summarizes detection histories and provides Cormack-Jolly-Seber
estimates with or without adjustment of potential tag-failure}
\usage{
cjs.fn(detect.in, L.in = NULL, seeds.in = NULL, se.out = F, d = NULL)
}
\arguments{
\item{detect.in}{Detection history in wide ATLAS format without rel.group or
bin columns}

\item{L.in}{L.in: Vector of mean probabilities tag is working given detection
at each site. If NULL, function will estimate the CJS parameters without tag
failure}

\item{seeds.in}{Optional starting estimates for CJS model parameters Si, pi,
li where i is 1:(total number of sites-1)}

\item{se.out}{(T|F) Should standard error of cjs parameters be estimated.
Uses variance matrix based on Hessian from MLE of cjs likelihood}

\item{d}{proportion of detected tags censored at a location}
}
\value{
Returns a two-column matrix with CJS estimates and standard errors
(if estimated)
}
\description{
CJS.fn summarizes the individual detection histories and
estimates seeds for Cormack-Jolly-Seber (CJS) likelhood function. It is based
on Skalski et al. (1998). This likelihood is also used to adjust survival for
estimated tag life as described in Townsend et al. (2006).  If no tag-life
data provided, only the unadjusted CJS survival will be returned.
}
