% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.fn.R
\name{correct.fn}
\alias{correct.fn}
\title{Restrains uncorrected probability estimates between 0 and 1}
\usage{
correct.fn(x)
}
\arguments{
\item{x}{Value to be rounded up/down to be between 0 to 1}
}
\value{
Returns original value if between 0 and 1, otherwise rounds up to 1e-10 or down to 1-1e-10
}
\description{
The Cormack-Jolly-Seber (CJS) likelihood estimates are currently bounded between 0 and
1, which may bias survival estimates low in cases of low detection probability.
This is not usually the case with active tag technology with typically high
detection rates. The CJS estimate can be greater than 100\%, and this may
be included in later iterations.
}
