% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.fn.R
\name{correct.fn}
\alias{correct.fn}
\title{Restrains uncorrected probability estimates between 0 and 1}
\usage{
correct.fn(x)
}
\arguments{
\item{x}{Value to be rounded up/down to be between 0 to 1}
}
\value{
Returns original value if between 0 and 1, otherwise rounds up to 1e-10 or down to 1-1e-10
}
\description{
Estimation of Cormack-Jolly-Seber (CJS) survival parameters can
be greater than 1 when detection probabilities are low. To reduce computational
failure when maximizing the likelihood, correct.fn bounds the CJS parameter
estimates between 0 and 1 in the likelihood.  This may result in survival
estimates biased low, but low detection rates do not usually occur in studies
with active tag technology.
}
