% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayhr.fn.R
\name{dayhr.fn}
\alias{dayhr.fn}
\title{Standardize date-times input}
\usage{
dayhr.fn(x, secs = T, AMPM = F)
}
\arguments{
\item{x}{character vector of date-time values
current accepted input formats: "m/\%d/\%Y \%I:\%M:\%S \%p"                                (AM/PM)
"\%Y-\%m-\%d \%H:\%M:\%S" or "\%m/\%d/\%Y \%H:\%M:\%S"    (with seconds)
"Y-\%m-\%d \%H:\%M" or "\%m/\%d/\%Y \%H:\%M"        (without seconds)}

\item{secs}{are seconds included in time values. Logical}

\item{AMPM}{is time format in AM, PM format. Logical}
}
\value{
striptime object in "\%Y-\%m-\%d \%H:\%M:\%S" or "\%Y-\%m-\%d \%H:\%M" format
}
\description{
Convert date-time into yyyy-mm-dd hh:mm:ss format
}
\examples{
\dontrun{
dayhr.fn("2021-05-10 14:22:34", secs =T , AMPM = F)
}
}
