% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic}
\alias{harmonic}
\title{Estimates harmonic mean}
\usage{
harmonic(t.in)
}
\arguments{
\item{t.in}{Numeric vector (no NAs).}
}
\value{
This function returns a character vector of harmonic mean and standard error
}
\description{
This function estimates harmonic mean.
}
