% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single.rel}
\alias{single.rel}
\title{Single Release ATLAS Practice File.}
\format{
These data are from a study of acoustic-tagged salmon using a release
above McNary Dam in 2014.  A data frame with 8 columns and 9996 rows in the
ATLAS file format:
\describe{
\item{V1}{Release group (ie. R1)}
\item{V2}{tag lot number (ie. 1,2,...)}
\item{V3}{tag id code}
\item{V4}{tag activation date/time  (yyyy-mm-dd hh:mm:ss)}
\item{V5}{release date/time  (yyyy-mm-dd hh:mm:ss)}
\item{V6}{detecton site designation}
\item{V7}{detection indicator (0: not detected; 1: detected; 2: detected & censored)}
\item{V8}{tag detection date/time  (yyyy-mm-dd hh:mm:ss) at that site}
...
}
}
\usage{
single.rel
}
\description{
Single Release ATLAS Practice File.
}
\references{
\href{https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=&cad=rja&uact=8&ved=2ahUKEwjv262c3Ln2AhW2JzQIHde_B40QFnoECAgQAQ&url=http\%3A\%2F\%2Fpweb.crohms.org\%2Ftmt\%2Fdocuments\%2FFPOM\%2F2010\%2FNWW\%2520Research\%2FCompliance\%2520Monitoring\%2520at\%2520McNary\%2520Dam\%2520in\%25202014_Final\%2520Report.pdf&usg=AOvVaw1LD6qSRn8oChimStypCy57}{McNary Dam survival study}
}
\keyword{datasets}
