% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taglife.data}
\alias{taglife.data}
\title{Taglife ATLAS Practice File.}
\format{
A data frame with 1 columns and 100 rows:
\describe{
\item{tag_life_days}{length of time from activation to tag failure in days}
...
}
}
\usage{
taglife.data
}
\description{
Taglife ATLAS Practice File.
}
\keyword{datasets}
