% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thist0.R
\name{thist0}
\alias{thist0}
\title{Creates a list with counts per history and a vertical matrix of observed histories only.}
\usage{
thist0(x)
}
\arguments{
\item{x}{Detection history in wide ATLAS format without rel.group or bin columns.}
}
\value{
This function returns a summary of counts.in per history and unique detection matrix in the same order.
}
\description{
Creates a list with counts per history and a vertical matrix of observed histories only.
}
